[RISCV] Add matching of codegen patterns to RISCV Bit Manipulation Zbbp asm instructions
Summary:
This patch provides optimization of bit manipulation operations by enabling the +experimental-b target feature.
It adds matching of single block patterns of instructions to specific bit-manip instructions belonging to both the permutation and the base subsets of the experimental B extension of RISC-V.
It adds also the correspondent codegen tests.

This patch is based on Clifford Wolf's proposal for the bit manipulation extension of RISCV:
https://github.com/riscv/riscv-bitmanip/blob/master/bitmanip-0.92.pdf

Reviewers: simoncook, edward-jones, asb, lewis-revill

Subscribers: hiraditya, rbar, johnrusso, sabuasal, niosHD, kito-cheng, shiva0217, MaskRay, zzheng, rogfer01, MartinMosbeck, brucehoult, the_o, rkruppe, PkmX, jocewei, psnobl, benna, Jim, lenary, s.egerton, pzheng, sameer.abuasal, apazos, luismarques, evandro, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D79873